refactor: extract smaller components
For improved readability and testability.
